Fix kernel message structure to match apps

+++ software/kernel/modules/dmni.c
@@ -14,9 +14,9 @@
 #include "dmni.h"
 #include "hal.h"
 
-void dmni_read(unsigned int *payload_address, unsigned short payload_size)
+void dmni_read(int *payload_address, int payload_size)
 {
-	HAL_DMNI_SIZE = payload_size;
+	HAL_DMNI_SIZE = (unsigned int)payload_size;
 	HAL_DMNI_OP = HAL_DMNI_WRITE;
 	HAL_DMNI_ADDRESS = (unsigned int)payload_address;
 	HAL_DMNI_START = 1;
